src/order-api/pyproject.toml
[tool.poetry]
name = "orderapi"
version = "0.1.0"
description = ""
authors = ["David <finnspalding@gmail.com>"]
readme = "../../README.md"

[tool.poetry.dependencies]
python = "^3.12"
tornado = "^6.4.2"
psycopg = {extras = ["binary"], version = "^3.2.3"}
pyyaml = "^6.0.2"
requests = "^2.32.3"
redis = "^5.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
